Moved plotting functions from common.py to plot.py
