+++ freud/environment.pyx
@@ -871,13 +871,14 @@ cdef class MatchEnv:
         return repr(self)
 
     def _repr_png_(self):
+        import plot
         try:
             count = np.unique(self.clusters, return_counts=True)
         except ValueError:
             return None
         if not self.m_box.is2D():
             return None
-        return freud.common.plot_clusters(count[0], count[1])
+        return plot.plot_clusters(count[0], count[1])
 
 
 cdef class AngularSeparation:
