+++ freud/voronoi.pyx
@@ -432,7 +432,8 @@ class Voronoi(object):
         return repr(self)
 
     def _repr_png_(self):
+        import plot
         if not self._box.is2D():
             return None
         else:
-            return freud.common.draw_voronoi(self._box, self.polytopes)
+            return plot.draw_voronoi(self._box, self.polytopes)
